+++ app/src/main/java/com/tutk/IOTC/Monitor.kt
@@ -708,7 +708,7 @@ class Monitor @JvmOverloads constructor(
 
     }
 
-    override fun receiveFrameData(camera: Camera?, avChannel: Int, bmp: Bitmap?, time: Int) {
+    override fun receiveFrameData(camera: Camera?, avChannel: Int, bmp: Bitmap?, time: Long) {
         if (avChannel != mAvChannel) {
             Liotc.d("Monitor", "receiveFrameData error [$avChannel],[$mAvChannel],[${bmp == null}]")
         }
