+++ app/src/main/java/com/tutk/IOTC/camera/AVChannel.kt
@@ -386,7 +386,7 @@ interface IAVChannelListener {
     fun onAVChannelReceiverFrameData(channel: Int, bitmap: Bitmap?)
 
     /**图片内容，包含时间戳*/
-    fun onAVChannelReceiverFrameData(channel: Int, bitmap: Bitmap?, time: Int)
+    fun onAVChannelReceiverFrameData(channel: Int, bitmap: Bitmap?, time: Long)
 
     fun onAVChannelReceiverFrameInfo(
         channel: Int,
