+++ app/src/main/java/com/tutk/IOTC/listener/IRegisterIOTCListener.kt
@@ -12,7 +12,7 @@ import com.tutk.IOTC.Camera
 interface OnFrameCallback {
     fun receiveFrameData(camera: Camera?, avChannel: Int, bmp: Bitmap?)
 
-    fun receiveFrameData(camera: Camera?,avChannel: Int,bmp: Bitmap?,time:Int)
+    fun receiveFrameData(camera: Camera?,avChannel: Int,bmp: Bitmap?,time:Long)
 
     fun receiveFrameInfo(
         camera: Camera?,
